:bug: fix: refactor Redis cache implementation to use UniversalClient and streamline method calls
